Handle data files which specify the catchement population.
This is currently done in a fairly inefficient way.
The catchment population (as a proportion of the total population) is passed to
C++ on each evaluation of the likelihood.
Firstly, this could lead to a data bottleneck.
Secondly, the total population must be specified in both C++ and Python which
has potential to lead to bugs (DRY).